requirements.txt
anyio==4.9.0
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==4.2.5
django-cors-headers==4.7.0
django-countries==7.6.1
django-pwa==2.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
googletrans==4.0.2
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
pillow==11.2.1
psycopg==3.2.9
psycopg2-binary==2.9.10
PyJWT==2.9.0
PyMySQL==1.1.1
qrcode==8.2
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.4.0
